+++ src/main/java/com/semgarcorp/ferreteriaSemGar/repositorio/CotizacionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CotizacionRepository extends JpaRepository<Cotizacion, Long> {
+public interface CotizacionRepository extends JpaRepository<Cotizacion, Integer> {
 }
