+++ src/main/java/com/semgarcorp/ferreteriaSemGar/repositorio/NominaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NominaRepository extends JpaRepository<Nomina, Long> {
+public interface NominaRepository extends JpaRepository<Nomina, Integer> {
 }
